requirements.txt
stable-baselines3
gymnasium
torch
numpy
pandas
matplotlib
requests
websocket-client
python-dotenv
